requirements.txt
anthropic>=0.40.0
datasets>=3.1.0
huggingface_hub>=0.27.1
openai>=1.59.7
pandas==1.4.1
python-dotenv>=1.0.1
